got rid of extra update hist and fixed rint issue

+++ src/environment/odor_plumes.py
@@ -75,8 +75,8 @@ class StaticGaussianRibbon:
         self.source_location: np.ndarray = plume_dict['SOURCE_LOCATION_MM']
         self.max_frames = plume_dict['STOP_FRAME']
         self.frame_number = 0
-        dim_x = np.rint(self.frame_x/self.mm_per_px)
-        dim_y = np.rint(self.frame_y/self.mm_per_px)
+        dim_x = np.rint(self.frame_x/self.mm_per_px).astype(int)
+        dim_y = np.rint(self.frame_y/self.mm_per_px).astype(int)
         y_list = self.mm_per_px*np.arange(0,dim_y)
         y_coords = np.tile(y_list, (dim_x,1))
         self.frame = self.source_conc*np.exp((-(y_coords-self.source_location[1])**2)/(2*self.sigma**2))
